Moved all the record extensions into Factory files
