+++ gary-m/src/PayerEDI.Data/Models/CommunicationContact.cs
@@ -1,0 +1,11 @@
+namespace PayerEDI.Data.Models;
+
+public record CommunicationNumber(string Number, CommunicationNumberQualifier Qualifier);
+
+public record CommunicationsContact(
+    string ContactFunctionCode,
+    string? Name,
+    CommunicationNumber? PrimaryNumber,
+    CommunicationNumber? SecondaryNumber,
+    CommunicationNumber? TertiaryNumber
+);
