+++ gary-m/src/PayerEDI.Data/Models/Factory/CommunicationNumber.Factory.cs
@@ -1,6 +1,34 @@
-namespace PayerEDI.Data.Models;
+using System.Reflection;
+using System.Runtime.Serialization;
 
-public record CommunicationNumber(string Number, CommunicationNumberQualifier Qualifier);
+namespace PayerEDI.Data.Models.Factory;
+
+public static class CommunicationNumberQualifierExtensions
+{
+    extension(CommunicationNumberQualifier)
+    {
+        public static CommunicationNumberQualifier? FromValue(string value)
+        {
+            if (!string.IsNullOrWhiteSpace(value))
+            {
+                foreach (var qualifier in Enum.GetValues<CommunicationNumberQualifier>())
+                {
+                    var member = typeof(CommunicationNumberQualifier).GetMember(
+                        qualifier.ToString()
+                    )[0];
+                    var enumMember = member.GetCustomAttribute<EnumMemberAttribute>();
+
+                    if (string.Equals(enumMember?.Value, value, StringComparison.OrdinalIgnoreCase))
+                    {
+                        return qualifier;
+                    }
+                }
+            }
+
+            return null;
+        }
+    }
+}
 
 public static class CommunicationNumberExtensions
 {
