+++ gary-m/tests/PayerEDI.Tests/Models/IndividualOrOrganization/NonPersonTests.cs
@@ -1,6 +1,7 @@
 using Bogus;
 using EdiFabric.Templates.Hipaa5010;
 using PayerEDI.Data.Models;
+using PayerEDI.Data.Models.Factory;
 
 namespace PayerEDI.Tests.Models.IndividualOrOrganization;
