+++ gary-m/tests/PayerEDI.Tests/Models/IndividualOrOrganization/PersonTests.cs
@@ -1,6 +1,7 @@
 using Bogus;
 using EdiFabric.Templates.Hipaa5010;
 using PayerEDI.Data.Models;
+using PayerEDI.Data.Models.Factory;
 using Person = PayerEDI.Data.Models.Person;
 
 namespace PayerEDI.Tests.Models.IndividualOrOrganization;
